+++ demos/vue-supabase-todolist/CHANGELOG.md
@@ -1,5 +1,13 @@
 # vue-supabase-todolist
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [3aaee03]
+  - @journeyapps/powersync-sdk-web@0.5.0
+  - @powersync/vue@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes